+++ web/modules/custom/qs_activity/src/Form/EventEditForm.php
@@ -336,8 +336,16 @@ class EventEditForm extends EventEditFormBase {
     $fields['field_venue_lat']     = $form_state->getValue('latitude');
     $fields['field_venue_long']    = $form_state->getValue('longitude');
 
+    $original_event = clone $event;
+
     // Update new event.
-    $this->eventManager->update($event, $start_at, $end_at, $fields);
+    $updated_event = $this->eventManager->update($event, $start_at, $end_at, $fields);
+
+    // Process this only if the date or time has change form the original event.
+    if ($original_event->field_start_at->date != $updated_event->field_start_at->date || $original_event->field_end_at->date != $updated_event->field_end_at->date) {
+      // Send mail to subscribers, activity organizer(s) & activity maintainers.
+      $this->eventManager->sendUpdated($original_event, $updated_event, $this->currentUser->getAccount());
+    }
 
     drupal_set_message($this->t('qs_activity.events.form.edit.success @event', [
       '@event' => $event->getTitle(),