+++ CHANGELOG.md
@@ -25,6 +25,12 @@
  - send mail to community managers when new user(s) request access to community #264 - via communityApply form
  - send mail to community managers when new user(s) request access to community #264 - via register form
  - set the modal-open wrapper to `position: fixed` to remove scroll inertia on Safari #408
+ - send mail to subscribers of events when event is deleted #381
+ - send mail to activity organizers when events is deleted #381
+ - send mail to activity maintainers when events is deleted #381
+ - send mail to subscribers of events when event is updated #381
+ - send mail to activity organizers when events is updated #381
+ - send mail to activity maintainers when events is updated #381
  - add the role in communities listing of supervisor dashboard #202 
  - flip the card back when using the browser back button #520
  - remove confirm state text in ajax submit buttons #224